Add data adapter
